Time: 0 ms (100%), Space: 43.3 MB (8.86%) - LeetHub

+++ 2177-find-three-consecutive-integers-that-sum-to-a-given-number/2177-find-three-consecutive-integers-that-sum-to-a-given-number.java
@@ -2,7 +2,7 @@ class Solution {
     public long[] sumOfThree(long num) {
         long x=0;
         if(num%3==0){
-            x=(long)num/3;
+            x=num/3;
             return new long[]{x-1,x,x+1};
         }
         return new long[]{};
